Mapping entities (#4)
* - Mapping entities
 - changing script names

* - Creationg Operation enum

+++ src/main/resources/db/migration/V2__create_operation_type.sql
@@ -1,5 +1,5 @@
 CREATE TABLE operation_type (
-    operation_type_id SMALLINT PRIMARY KEY,
+    operation_type_id INTEGER PRIMARY KEY,
     description VARCHAR(50) NOT NULL UNIQUE
 );
