+++ src/main/resources/db/migration/V4__create_account_transaction.sql
@@ -1,7 +1,7 @@
 CREATE TABLE account_transaction (
     transaction_id BIGSERIAL PRIMARY KEY,
     account_id BIGINT NOT NULL,
-    operation_type_id SMALLINT NOT NULL,
+    operation_type_id INTEGER NOT NULL,
     amount NUMERIC(15,2) NOT NULL,
     event_date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
